Reorganized folders in messaging module
- Receivers/Transmitters now grouped under transceivers
- Messages grouped under messages (events+commands)
