+++ src/SchedulerSubsystem/Scheduler.java
@@ -1,14 +1,14 @@
 package SchedulerSubsystem;
 
 import ElevatorSubsytem.ElevatorUtilities;
-import Messaging.Commands.MoveElevatorCommand;
-import Messaging.Direction;
-import Messaging.Events.DestinationEvent;
-import Messaging.Events.ElevatorStateEvent;
-import Messaging.Events.FloorRequestEvent;
-import Messaging.Receivers.DMA_Receiver;
-import Messaging.SystemMessage;
-import Messaging.Transmitters.DMA_Transmitter;
+import Messaging.Messages.Commands.MoveElevatorCommand;
+import Messaging.Messages.Direction;
+import Messaging.Messages.Events.DestinationEvent;
+import Messaging.Messages.Events.ElevatorStateEvent;
+import Messaging.Messages.Events.FloorRequestEvent;
+import Messaging.Transceivers.Receivers.DMA_Receiver;
+import Messaging.Messages.SystemMessage;
+import Messaging.Transceivers.Transmitters.DMA_Transmitter;
 
 import com.sun.jdi.InvalidTypeException;
 import java.util.ArrayList;
