+++ test/unit/Networking/DMA_TransmitterTest.java
@@ -1,10 +1,10 @@
 package unit.Networking;
 
-import Messaging.Direction;
-import Messaging.Events.DestinationEvent;
-import Messaging.Events.SystemEvent;
-import Messaging.Receivers.DMA_Receiver;
-import Messaging.Transmitters.DMA_Transmitter;
+import Messaging.Messages.Direction;
+import Messaging.Messages.Events.DestinationEvent;
+import Messaging.Messages.Events.SystemEvent;
+import Messaging.Transceivers.Receivers.DMA_Receiver;
+import Messaging.Transceivers.Transmitters.DMA_Transmitter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
